Clean up Web class
Web class is now an instance, not static.

Inject Web class with a Request Handler. This means we can now mock out
how Web talks to Steam (for unit testing later).

Reduce embedded classes (Web class was getting too large).
